--- a/scripts/generate-data.js
+++ b/scripts/generate-data.js
@@ -2,7 +2,6 @@
 
 const resources = require('../data/resources.js');
 const generateData = require('../index.js');
-const utils = require('../scripts/utils.js');
 
 // -----------------------------------------------------------------------------
 
@@ -28,7 +27,9 @@ const complicatedWorkThatTakesTime = (resource, callback) => {
 		console.log('[%s] Worker %d \u2192 Unicode v%s',
 			getTime(), cluster.worker.id, version);
 
+		console.groupCollapsed();
 		generateData(version);
+		console.groupEnd();
 
 		complicatedWorkThatTakesTime(
 			resource.slice(1),
@@ -40,7 +41,7 @@ const complicatedWorkThatTakesTime = (resource, callback) => {
 	}
 };
 
-if (cluster.isMaster) {
+if (cluster.isPrimary) {
 
 	for (let index = 0; index < numCPUs; index++) {
 		cluster.fork();