@mathiasbynens 2024-11-08
Rebuilding all packages after merging this, I noticed that `Bidi_Mirroring_Glyph` no longer gets mentioned in the generated `README`s, e.g. https://github.com/node-unicode/unicode-16.0.0/commit/97cb5e34784799c100aa49f1a8ea626d884bcc49#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5. Is this intended?